--- a/worlds/pokemon_frlg/rules.py
+++ b/worlds/pokemon_frlg/rules.py
@@ -161,7 +161,7 @@ class PokemonFRLGLogic:
 
     def can_evolve(self, state: CollectionState, pokemon: str, item_names: tuple[str, ...]) -> bool:
         evo_data = data.evolutions[pokemon]
-        if state.has_any(item_names, self.player) and evo_data.method in self.evo_methods_required:
+        if state.has_any(item_names, self.player):
             if evo_data.method in EVO_METHODS_ITEM:
                 return state.has(self.world_item_id_map[evo_data.param], self.player)
             elif evo_data.method in EVO_METHODS_HELD_ITEM: