## What is this fixing or adding?
This is a whole bunch of small performance improvements.

Most commits have a small description going into more detail.

There is a lot of replacing the creating of new sets/lists within access rules with precomputed lists/tuples or literal tuples (tuples containing only literal values). Creating new sets/lists/non-literal tuples in access rules is generally a waste of cpu time because those sets/lists/tuples take time to create are only going to be immediately discarded after use. While python can automatically replace `for x in [1, 2, 3]:` with `for x in (1, 2, 3):`, it cannot do this replacement when passing `[1, 2, 3]` to another function, so `(1, 2, 3)` must be used explicitly.

There are more things that are now pre-calculated, such as the event items that oak's aide and dexsanity locations check for are now reduced to only the event items that exist in the world.

A bigger change is implementing `collect`/`remove` for logically being able to teach HMs (having logical access to pokemon that can learn HMs). About 90% of the collected advancements during a generation are pokemon events, and there are a lot of them, so care should be taken when working with FRLG's `collect` because it can quickly become worse performance compared to more expensive access rules that get called more often. If there is one change in this PR to be most unsure about, it would be this one.

I tried using `collect`/`remove` for oak's aide logic, but the extra code of having to count unique pokemon across wild/static/evolved resulted in `collect`/`remove` being slower for oak's aide logic, and that was before I made any changes to increase the performance of the existing oak's aide logic.

There is also a trivial bug fix to a couple of calls to `state.can_reach()` where the second, unused, argument was incorrect (it was the player ID when the second argument is the string hint). Because the argument was unused, it never had any effect on generation despite being incorrect.

## How was this tested?
I have generator modifications that output a hash of all placements so I generate a fixed seed before these changes and that same fixed seed with these changes and use that hash to determine that identical generation took place.

I also generated 500 3-player FRLG worlds with the fuzzer and the only failures were due to an extra test I run that all locations in Full accessibility worlds must be reachable (`Water Labyrinth - Gentleman Info` occasionally fails this for some reason, but this also occurs without the changes in this PR).

The 3 yamls I was using to generate (`python -O .\Generate.py --skip_output --seed 1` with progression balancing disabled through a meta yaml) with went from around 5.6s generation before to around 4.9s generation after with these changes, and produced the same output placements.

Generation of 10 template FRLG yamls (`python -O .\Generate.py --seed 1` with no playthrough) goes from about 7.5s to 7.0s for me.