@Mysteryem 2025-05-05
It was not clear in the commit messages giving durations of the Oak's aide locations, but those are from the `test\benchmark\locations.py` benchmark (with `gc.freeze()` and `gc.unfreeze()` added locally to stop `gc.collect()` being the majority of the runtime for almost every location test), which are per 100_000 calls.

Before:
```
Pokemon FireRed and LeafGreen took 0.6977 seconds per location in empty_state and 0.0867 in all_state. (all times summed for 100000 runs.)
Top times in empty_state:
  35.0280 in Route 10 Pokemon Center 1F - Oak's Aide Gift (Pokedex Progress)
  33.6589 in Route 2 Gate - Oak's Aide Gift (Pokedex Progress)
  33.5869 in Route 11 Gate 2F - Oak's Aide Gift (Pokedex Progress)
  33.5346 in Route 15 Gate 2F - Oak's Aide Gift (Pokedex Progress)
  33.4881 in Route 16 Gate 2F - Oak's Aide Gift (Pokedex Progress)
Top times in all_state:
  2.3643 in Route 15 Gate 2F - Oak's Aide Gift (Pokedex Progress)
  2.0555 in Route 16 Gate 2F - Oak's Aide Gift (Pokedex Progress)
  1.7854 in Route 11 Gate 2F - Oak's Aide Gift (Pokedex Progress)
  1.3825 in Route 10 Pokemon Center 1F - Oak's Aide Gift (Pokedex Progress)
  1.0594 in Route 2 Gate - Oak's Aide Gift (Pokedex Progress)
```

PR:
```
Pokemon FireRed and LeafGreen took 0.0937 seconds per location in empty_state and 0.0164 in all_state. (all times summed for 100000 runs.)
Top times in empty_state:
  4.1804 in Route 10 Pokemon Center 1F - Oak's Aide Gift (Pokedex Progress)
  4.1800 in Route 16 Gate 2F - Oak's Aide Gift (Pokedex Progress)
  4.1747 in Route 11 Gate 2F - Oak's Aide Gift (Pokedex Progress)
  4.1710 in Route 15 Gate 2F - Oak's Aide Gift (Pokedex Progress)
  4.1627 in Route 2 Gate - Oak's Aide Gift (Pokedex Progress)
Top times in all_state:
  0.4320 in Route 15 Gate 2F - Oak's Aide Gift (Pokedex Progress)
  0.3517 in Route 16 Gate 2F - Oak's Aide Gift (Pokedex Progress)
  0.2522 in Route 11 Gate 2F - Oak's Aide Gift (Pokedex Progress)
  0.1778 in Route 10 Pokemon Center 1F - Oak's Aide Gift (Pokedex Progress)
  0.1041 in Mt. Ember Exterior - Item Near Summit
```