@vyneras 2025-05-07
@Mysteryem You have some comment about a few other optimizations that could be made. Are any of those things you want to add to this PR or should I go ahead and merge?